+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unrelease
 [Compare](https://github.com/leoture/MockSwift/compare/v0.2.0...HEAD)
 #### Added
+- Add Predicate .match(description:keyPath:) [#38](https://github.com/leoture/MockSwift/pull/38)
 - Add willThrow [#37](https://github.com/leoture/MockSwift/pull/37)
 - Add willReturn with a List [#36](https://github.com/leoture/MockSwift/pull/36)
 - Add Predicates with Comparables [#28](https://github.com/leoture/MockSwift/pull/28)